% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Shortest path Algorithm}
\usage{
dijkstra(graph, init.node )
}
\arguments{
\item{graph}{data frame with three variables \code{(v1,v2,w)} and variables should be edges of the graph with the weight of edge}

\item{init.node}{numeric scale exist in the graph}
}
\value{
shortest distance vector from \code{init.node} to each nodes in \code{graph}
}
\description{
Return the shortest path from \code{init.node} to each nodes in a \code{graph}, using weight vector in graph data
         frame find shortest distance and return single vector as a result with distance to each nodes.
}
\examples{
 sample_graph <-  data.frame(
  v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
  v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
  w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)
)

 dijkstra(sample_graph, 1)

 dijkstra(sample_graph, 3)
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra27s_algorithm}
}
