% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find greatest common divisor (GCD) of two integers (numbers).}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{is a numerical number}

\item{b}{is a numerical number}
}
\value{
greatest common divisor from integers \code{a} and \code{b}
}
\description{
Find greatest common divisor (GCD) of two integers (numbers).
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\author{
Chandika
}
