% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{This is the graph data from wiki data}
\description{
This is the graph data from wiki data
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra27s_algorithm}
}
\keyword{data}
